fix : 결제관련 insert MyBatis -> JPA

+++ src/main/java/com/dogdailylog/payment/dao/PaymentMapper.java
@@ -8,16 +8,16 @@ import com.dogdailylog.payment.model.PaymentInfo;
 @Repository
 public interface PaymentMapper {
 	public int insertPayment(
-			@Param("bookingId") int bookingId
+			@Param("bookingId") Long bookingId
 			, @Param("userId") int userId
 			, @Param("price") int price
 			, @Param("payment") String payment);
 
-	public PaymentInfo selectLastPaymentByBookingId(int id);
+	public PaymentInfo selectLastPaymentByBookingId(Long id);
 	
 	public int updatePaymentById(
 			@Param("id") int id
 			, @Param("approval") int approval);
 
-	public int deletePaymentByBookingId(int bookingId);
+	public int deletePaymentByBookingId(Long bookingId);
 }
